- Switched to INI configuration file (PHP is a bad format for configuration, too slow and too large a temptation to make it too dynamic) - Improved exception handling - Added error reporting to mail - Improved default routing - Improved PHP settings bootstrapping - Refactored User Directory - Worked on user provisioning - Improved bare bones MVC
git-svn-id: https://svn.surfnet.nl/svn/coin-eb/engineblock/trunk@147 528793b8-127e-4309-9901-b74759c7208d

+++ library/EngineBlock/Corto/Module/Services.php
@@ -66,10 +66,12 @@ class EngineBlock_Corto_Module_Services extends Corto_Module_Services
         }
 
         $xml = Corto_XmlToArray::array2xml($entitiesDescriptor, 'md:EntitiesDescriptor', true);
-        if ($this->_server->getConfig('debug', false) && ini_get('allow_url_fopen')) {
+
+        $schemaUrl = 'http://docs.oasis-open.org/security/saml/v2.0/saml-schema-metadata-2.0.xsd';
+        if ($this->_server->getConfig('debug', false) && ini_get('allow_url_fopen') && file_exists($schemaUrl)) {
             $dom = new DOMDocument();
             $dom->loadXML($xml);
-            if (!$dom->schemaValidate('http://docs.oasis-open.org/security/saml/v2.0/saml-schema-metadata-2.0.xsd')) {
+            if (!$dom->schemaValidate($schemaUrl)) {
                 echo '<pre>'.htmlentities(Corto_XmlToArray::formatXml($xml)).'</pre>';
                 throw new Exception('Metadata XML doesnt validate against XSD at Oasis-open.org?!');
             }
